fix(db): regenerate migrations for _linchkit schema + fix test fixtures
- Delete old migration (public._linchkit_* tables)
- Generate fresh migration with _linchkit PG schema tables
- Fix outbox-worker test: add missing idx_events_tenant index
- Fix drizzle-data-provider test: amount column type numeric (not double precision)

+++ packages/core/__tests__/drizzle-data-provider.integration.test.ts
@@ -90,7 +90,7 @@ describe.skipIf(!dbAvailable)("DrizzleDataProvider (integration)", () => {
         "_version" integer DEFAULT 1 NOT NULL,
         "deleted_at" timestamp,
         "title" varchar(255) NOT NULL,
-        "amount" double precision,
+        "amount" numeric,
         "status" varchar(50)
       )
     `));
